+++ init.el
@@ -67,10 +67,10 @@
   (load (concat modules-dir "22-zettelkasten.el")) ; Folgezettel sequences (C-c n z)
   (load (concat modules-dir "23-fixed-tabs.el"))   ; Route commands to named tabs
   (load (concat modules-dir "24-readwise.el"))     ; Readwise import (C-c r s)
-  (load (concat modules-dir "27-denote-identifiers.el")) ; Identifier integrity for checking after md files moved to org format
-  (load (concat modules-dir "28-writing-projects.el")) ; Writing projects (C-c n p)
   (load (concat modules-dir "25-inbox-review.el")) ; Inbox review for old notes from Obsidian
   (load (concat modules-dir "26-performance.el")) ; Fixing some performance issue after adding 3,5k files to the silos.
+  (load (concat modules-dir "27-denote-identifiers.el")) ; Identifier integrity for checking after md files moved to org format
+  (load (concat modules-dir "28-writing-projects.el")) ; Writing projects (C-c n p)
   )
 
 (add-hook 'emacs-startup-hook